# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "call4me API",
    "version": "1.1.0",
    "description": "Manage the account's phone numbers and retrieve existing call recordings. Phone calls and other account tools are available through MCP."
  },
  "servers": [
    {
      "url": "https://call4.me"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A call4me API key or OAuth access token for the site or MCP resource."
      }
    }
  },
  "paths": {
...
    "/api/calls/{call_id}/recordings": {
      "get": {
        "operationId": "getCallRecordings",
        "description": "Get existing recordings of a finished call owned by the signed in account. Returns fresh download links when available. An empty list means no recording is available yet. Links may expire; request this again for fresh links. Only share the links with the user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 1,
              "maxLength": 40,
              "description": "The call id returned when placing or listing calls."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available recordings and fresh download links",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "call_id": {
                      "type": "string"
                    },
                    "recordings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "download_urls": {
                            "type": "object",
                            "properties": {
                              "mp3": {
                                "type": "string",
                                "format": "uri"
                              },
                              "wav": {
                                "type": "string",
                                "format": "uri"
                              }
                            },
                            "additionalProperties": false
                          },
                          "duration_millis": {
                            "anyOf": [
                              {
                                "type": "number",
                                "minimum": 0
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "started_at": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "ended_at": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "download_urls",
                          "duration_millis",
                          "started_at",
                          "ended_at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "call_id",
                    "recordings",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}